Optimize data types as much as possible for binary reading.

+++ lib/src/main/java/org/sehkah/doon/tools/extractor/lib/logic/entity/base/LandInfo.java
@@ -5,6 +5,6 @@ import java.util.List;
 public record LandInfo(
         long LandId,
         boolean IsDispNews,
-        long GameMode,
+        int GameMode,
         List<Long> AreaIds) {
 }
